refactor(index): split cow_btree into a module directory (8 files)
Pure relocation of the 3059-line cow_btree.rs into cow_btree/, by layer:
  mod.rs                  CowBTree<L,B> + byte primitives (read_u64/read_width/FIELD/STRIDE)
  leaf/mod.rs             Leaf<L> enum + dispatch, LeafFormat, LeafInsert, cross-format helpers
  leaf/aos.rs             AosLeaf
  leaf/compact.rs         compact on-disk format (offsets/CompactLayout/header/packing_fits) + CompactLeaf
  leaf/compact_indexed.rs CompactIndexedLeaf + gallop_lower_bound (extends compact)
  node.rs                 Node/Branch + tree algorithms, Split/Combined, make_mut_gate
  cursor.rs               RangeIter
  tests.rs                the test module

No logic change. Each shared helper sits in the lowest module covering its
callers; the only new code is compact::is_indexed so from_parts needn't read
the raw offsets. LeafFormat/RangeIter stay pub (unchanged public API). 22
tests pass; clippy --all-targets + fmt clean.
